requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pyyaml==6.0.1
aiohttp==3.9.1
psutil==5.9.6
feedparser==6.0.10
pytz==2023.3.post1